+++ src/main/kotlin/binding/KanamaResourceFormatLoader.kt
@@ -133,7 +133,7 @@ object KanamaResourceFormatLoader {
       )
 
     val constructObject =
-      GodotFFI.lookup("classdb_construct_object2", FunctionDescriptor.of(ADDRESS, ADDRESS))
+      GodotFFI.lookup("classdb_construct_object3", FunctionDescriptor.of(ADDRESS, ADDRESS))
     val objectSetInstance =
       GodotFFI.lookup("object_set_instance", FunctionDescriptor.ofVoid(ADDRESS, ADDRESS, ADDRESS))
     val obj = constructObject.invoke(cls.parentName) as MemorySegment
@@ -198,7 +198,7 @@ object KanamaResourceFormatLoader {
   @JvmStatic
   fun createInstance(userdata: MemorySegment, notifyPostinitialize: Byte): MemorySegment {
     val constructParent =
-      GodotFFI.lookup("classdb_construct_object2", FunctionDescriptor.of(ADDRESS, ADDRESS))
+      GodotFFI.lookup("classdb_construct_object3", FunctionDescriptor.of(ADDRESS, ADDRESS))
     val objectSetInstance =
       GodotFFI.lookup("object_set_instance", FunctionDescriptor.ofVoid(ADDRESS, ADDRESS, ADDRESS))
     val obj = constructParent.invoke(cls.parentName) as MemorySegment